+++ backend/views/layouts/sidebar.php
@@ -194,7 +194,7 @@ use yii\helpers\Url;
                         [
                             'label' => 'Расходы',
                             'icon'  => 'fas fa-dollar-sign',
-                            'iconClassAdded' => 'text-danger',
+                            'iconClassAdded' => 'text-success',
                             'items' => [
                                 [
                                     'label' => 'Категории расходов',
@@ -207,11 +207,21 @@ use yii\helpers\Url;
                                     'icon'  => 'far fa-plus-square',
                                     'url'   => ['/expenses/create'],
                                 ],
+                                [
+                                    'label' => 'Список затрат',
+                                    'icon'  => 'fas fa-list',
+                                    'url'
+                                            => ['/expenseslist/index'],
+                                ],
+                                [
+                                    'label' => 'Добавить затраты',
+                                    'icon'  => 'far fa-plus-square',
+                                    'url'   => ['/expenseslist/create'],
+                                ],
                             ]
                         ],
+                        ['label' => 'Статистика', 'icon' => 'fas fa-chart-pie', 'url' => ['/statistic'],'iconClassAdded' =>'text-info'],
                         ['label' => 'Заметки', 'icon' => 'fa fa-list-ul', 'url' => ['/todo/todo/index'],'iconClassAdded' =>'text-info'],
-
-
                         /*['label' => 'LABELS', 'header' => true],
                         ['label' => 'Important', 'iconStyle' => 'far', 'iconClassAdded' => 'text-danger'],
                         ['label' => 'Warning', 'iconClass' => 'nav-icon far fa-circle text-warning'],
