+++ common/modules/calendar/views/event/_expenseslist.php
@@ -1,0 +1,2 @@
+<?php
+echo __DIR__;
